added max(abs(eigenvalue)) in eigenValue calculator, added fine handling of lambdaRho. finally response seems ok, still different from matlab response...But at least seems coherent

+++ jerman.cpp
@@ -62,7 +62,7 @@ int main( int argc, char* argv[] )
     float tau = vm["tau"].as<float>();
 
     constexpr unsigned int Dimension = 3;
-    using PixelType = float;
+    using PixelType = double;
     using ImageType = itk::Image< PixelType, Dimension >;
 
     typedef itk::ImageFileReader<ImageType> ReaderType;
